requirements.txt
streamlit
requests
crewai
langchain_community
duckduckgo-search
python-dotenv
langchain_google_genai